smart_speaker/requirement.txt
silero-vad
sounddevice
soundfile
torch
torchaudio
SpeechRecognition
gTTS
groq
python-dotenv
pydub
streamlit
streamlit-webrtc 
av